+++ packages/international-phone-input/src/docs/Component.stories.tsx
@@ -47,6 +47,7 @@ export const international_phone_input: Story = {
                         true,
                     )}
                     clear={boolean('clear', false)}
+                    success={boolean('success', false)}
                 />
             </div>
         );