+++ packages/pure-cell/src/components/text/component.tsx
@@ -1,10 +1,10 @@
-import React, { ReactNode, useContext } from 'react';
+import React, { type ReactNode, useContext } from 'react';
 import cn from 'classnames';
 
 import { getDataTestId } from '@alfalab/core-components-shared';
 import { type Color, Text as TypographyText } from '@alfalab/core-components-typography';
 
-import { PureCellContext } from '../../component';
+import { PureCellContext } from '../../context';
 
 import styles from './index.module.css';