+++ packages/universal-modal/src/components/content/desktop/Component.desktop.tsx
@@ -1,17 +1,17 @@
-import React, { FC, useContext } from 'react';
+import React, { type FC, useContext } from 'react';
 import cn from 'classnames';
 
 import { UniversalModalContext } from '../../../context/universal-modal-context';
-import { BaseContent, ContentProps } from '../base-content/base-content';
+import { BaseContent, type ContentProps } from '../base-content/base-content';
 
 import styles from './desktop.module.css';
 
-export type ContentDesktopProps = ContentProps & {
+export interface ContentDesktopProps extends ContentProps {
     /**
      * Размер (только для desktop версии компонента)
      */
     size?: 's' | 500;
-};
+}
 
 export const ContentDesktop: FC<ContentDesktopProps> = ({ className, ...restProps }) => {
     const { hasHeader, hasFooter } = useContext(UniversalModalContext);