+++ packages/root/CHANGELOG.md
@@ -1,5 +1,89 @@
 # @alfalab/core-components
 
+## 49.3.0
+
+### Minor Changes
+
+<sup><time>05.09.2025</time></sup>
+
+#### [#1863](https://github.com/core-ds/core-components/pull/1863)
+
+-   Добавлен обработчик ошибки загрузки изображения "onError"
+-   Исправлено описание свойства "alt"
+
+<sup><time>05.09.2025</time></sup>
+
+#### [#1813](https://github.com/core-ds/core-components/pull/1813)
+
+-   Добавлена возможность прокидывать restProps на root для компонентов RadioGroup, Space, Tab и Tabs
+
+### Patch Changes
+
+<sup><time>05.09.2025</time></sup>
+
+#### [#1831](https://github.com/core-ds/core-components/pull/1831)
+
+-   Исправлена проблема со специфичностью в компоненте `PopupSheet`. Увеличена специфичность селектора `.component` для корректного применения стилей позиционирования.
+
+<sup><time>05.09.2025</time></sup>
+
+#### [#1805](https://github.com/core-ds/core-components/pull/1805)
+
+-   Переменные типографики вынесены в отдельный файл `typography-vars.css`
+
+<sup><time>05.09.2025</time></sup>
+
+#### [#1831](https://github.com/core-ds/core-components/pull/1831)
+
+-   Исправлена проблема со специфичностью в компоненте `Tooltip`
+
+-   Обновлены зависимости
+    -   @alfalab/core-components-card-image@4.1.0
+    -   @alfalab/core-components-popup-sheet@2.0.3
+    -   @alfalab/core-components-themes@14.1.2
+    -   @alfalab/core-components-tooltip@8.0.2
+    -   @alfalab/core-components-radio-group@5.2.0
+    -   @alfalab/core-components-space@4.1.0
+    -   @alfalab/core-components-tabs@9.1.0
+    -   @alfalab/core-components-gallery@6.0.3
+
+## 49.2.0
+
+### Minor Changes
+
+<sup><time>29.08.2025</time></sup>
+
+#### [#1833](https://github.com/core-ds/core-components/pull/1833)
+
+-   Добавлен пропс `transformLinkUri` позволяющий отключить трансформацию для ссылок с неизвестными форматами
+
+### Patch Changes
+
+<sup><time>29.08.2025</time></sup>
+
+#### [#1838](https://github.com/core-ds/core-components/pull/1838)
+
+-   Исправлена типизация. Типы переведены на интерфейсы.
+
+<sup><time>29.08.2025</time></sup>
+
+#### [#1842](https://github.com/core-ds/core-components/pull/1842)
+
+-   Исправлена темизация `site` для `FormControl` и `SliderInput`
+
+<sup><time>29.08.2025</time></sup>
+
+#### [#1842](https://github.com/core-ds/core-components/pull/1842)
+
+-   Исправлена специфичность слайдера
+
+-   Обновлены зависимости
+    -   @alfalab/core-components-markdown@2.1.0
+    -   @alfalab/core-components-universal-modal@2.0.2
+    -   @alfalab/core-components-popup-sheet@2.0.2
+    -   @alfalab/core-components-themes@14.1.1
+    -   @alfalab/core-components-slider-input@10.0.2
+
 ## 49.1.0
 
 ### Minor Changes